+++ android/capacitor/src/main/java/com/getcapacitor/WebViewLocalServer.java
@@ -342,40 +342,29 @@ public class WebViewLocalServer {
                     isHtmlText = true;
                     break;
                 }
-            }
-            if (isHtmlText) {
-                HttpURLConnection conn = (HttpURLConnection) new URL(url).openConnection();
-                for (Map.Entry<String, String> header : headers.entrySet()) {
-                    conn.setRequestProperty(header.getKey(), header.getValue());
-                }
-                String getCookie = CookieManager.getInstance().getCookie(url);
-                if (getCookie != null) {
-                    conn.setRequestProperty("Cookie", getCookie);
-                }
-                conn.setRequestMethod(method);
-                conn.setReadTimeout(30 * 1000);
-                conn.setConnectTimeout(30 * 1000);
-                if (request.getUrl().getUserInfo() != null) {
-                    byte[] userInfoBytes = request.getUrl().getUserInfo().getBytes(StandardCharsets.UTF_8);
-                    String base64 = Base64.encodeToString(userInfoBytes, Base64.NO_WRAP);
-                    conn.setRequestProperty("Authorization", "Basic " + base64);
-                }
-                /**
-                    * This should be the final step in the setup process of the HttpURLConnection object.
-                    *
-                    * Initiate the connection with follow redirects disable. It allows us
-                    * identifying any redirection response and avoid intercepting such request.
-                    * This is because returning a 200 response when a redirection response (ie. 301 or 302)
-                    * is expected could result in issues with resolving relative URLs for static assets
-                    * in the HTML.
-                    */
-                conn.setInstanceFollowRedirects(false);
-                conn.connect();
+                if (isHtmlText) {
+                    HttpURLConnection conn = (HttpURLConnection) new URL(url).openConnection();
+                    for (Map.Entry<String, String> header : headers.entrySet()) {
+                        conn.setRequestProperty(header.getKey(), header.getValue());
+                    }
+                    String getCookie = CookieManager.getInstance().getCookie(url);
+                    if (getCookie != null) {
+                        conn.setRequestProperty("Cookie", getCookie);
+                    }
+                    conn.setRequestMethod(method);
+                    conn.setReadTimeout(30 * 1000);
+                    conn.setConnectTimeout(30 * 1000);
+                    if (request.getUrl().getUserInfo() != null) {
+                        byte[] userInfoBytes = request.getUrl().getUserInfo().getBytes(StandardCharsets.UTF_8);
+                        String base64 = Base64.encodeToString(userInfoBytes, Base64.NO_WRAP);
+                        conn.setRequestProperty("Authorization", "Basic " + base64);
+                    }
 
-                List<String> cookies = conn.getHeaderFields().get("Set-Cookie");
-                if (cookies != null && !cookies.isEmpty()) {
-                    for (String cookie : cookies) {
-                        CookieManager.getInstance().setCookie(url, cookie);
+                    List<String> cookies = conn.getHeaderFields().get("Set-Cookie");
+                    if (cookies != null) {
+                        for (String cookie : cookies) {
+                            CookieManager.getInstance().setCookie(url, cookie);
+                        }
                     }
                     CookieManager.getInstance().flush();
                 }